+++ backend/src/workspace/SysChecksFolder.class.php
@@ -80,15 +80,16 @@ class SysChecksFolder extends Workspace {
   }
 
   // TODO unit test
-  public function deleteSysCheckReports(array $checkIds): array {
+  public function deleteSysCheckReports(array $checkIds): FileDeletionReport {
     $reports = $this->collectSysCheckReports($checkIds);
 
-    $deletionReport = [];
+    $deletionReport = new FileDeletionReport();
 
     foreach ($reports as $report) {
       /* @var SysCheckReportFile $report */
       $fullPath = "$this->workspacePath/SysCheck/reports/{$report->getFileName()}";
-      $deletionReport[$this->deleteFileFromFs($fullPath)][] = $report->getCheckId();
+      $fieldName = $this->deleteFileFromFs($fullPath);
+      $deletionReport->$fieldName[] = $report->getCheckId();
     }
 
     return $deletionReport;